#4 - Move migration runner logic to common and remove its dependencies on server code
