--- server/database/migration/migration.go
@@ -1,11 +1,0 @@
-package migration
-
-type Migration interface {
-	GetTimestamp() string
-	Up() error
-	Down() error
-}
-
-type MigrationRepository interface {
-	GetMigrations() []Migration
-}
